refactor fetch data from api
