+++ src/containers/PokemonList.js
@@ -1,9 +1,9 @@
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import React, { useEffect } from 'react';
-import CategoryFilter from './Category';
+import CategoryFilter from './CategoryFilter';
 import { changeFilter, addPokemon, categoryPokemon } from '../actions/actions';
-import PokemonAsList from './PokemonAsList';
+import PokemonAsList from './PokemonBlock';
 import { allPokemons, pokemonByCategory } from '../api-requests/request';
 import style from '../styles/PokemonList.module.css';
