+++ src/routes/Routes.js
@@ -9,7 +9,6 @@ const Routes = () => (
     <Switch>
       <Route exact path="/" component={Home} />
       <Route exact path="/pokemon/:pokeName" component={Pokemon} />
-      {/* path="/users/:userId" */}
       <Route exact path="/about" component={About} />
     </Switch>
   </BrowserRouter>
